move all binding types to separate file
